renaming to geometry service; restructuring
