+++ spec/services/gtfs_graph_spec.rb
@@ -409,7 +409,7 @@ describe GTFSGraph do
         # here we take advantage of the fact that there can be slight, allowable discrepancies between shape_dist_traveled and our algorithm.
         # It's possible the shape_dist_traveled given for the stop doesn't match the distance computed for the closest point
         # just want to make sure we're using the shape_dist_traveled ultimately
-        expect(RouteStopPattern.find_by_onestop_id!(rsp.onestop_id).stop_distances).not_to match_array(DistanceCalculation::calculate_distances(rsp))
+        expect(RouteStopPattern.find_by_onestop_id!(rsp.onestop_id).stop_distances).not_to match_array(Geometry::DistanceCalculation::calculate_distances(rsp))
       end
     end
